--- src/test/groovy/com/cedarsoftware/ncube/OddComparator.groovy
@@ -1,7 +1,0 @@
-package com.cedarsoftware.ncube
-
-class OddComparator implements Comparable {
-    int compareTo(Object o) {
-        return 0
-    }
-}
