--- src/test/groovy/com/cedarsoftware/ncube/TestCIStringAxisSingle.groovy
@@ -1,7 +1,0 @@
-package com.cedarsoftware.ncube
-
-import groovy.transform.CompileStatic
-
-@CompileStatic
-class TestCIStringAxisSingle extends TestCIStringAxis implements StringAxisTestsSingle {
-}
