--- src/test/groovy/com/cedarsoftware/ncube/TestStringAxisSingle.groovy
@@ -1,8 +1,0 @@
-package com.cedarsoftware.ncube
-
-import groovy.transform.CompileStatic
-
-@CompileStatic
-class TestStringAxisSingle extends TestStringAxis implements StringAxisTestsSingle {
-
-}
